+++ pkg/client/model_get_schemas_response.go
@@ -22,16 +22,26 @@ package client
 type GetSchemasResponse struct {
 	Resources    *[]ScimSchema `json:"Resources,omitempty"`
 	ItemsPerPage *int32        `json:"itemsPerPage,omitempty"`
-	Schemas      *[]string     `json:"schemas,omitempty"`
+	Schemas      []string      `json:"schemas"`
 	StartIndex   *int32        `json:"startIndex,omitempty"`
-	TotalResults *int32        `json:"totalResults,omitempty"`
+	TotalResults int32         `json:"totalResults"`
 }
 
 // NewGetSchemasResponse instantiates a new GetSchemasResponse object.
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewGetSchemasResponse() *GetSchemasResponse {
+func NewGetSchemasResponse(schemas []string, totalResults int32) *GetSchemasResponse {
+	p := GetSchemasResponse{}
+	p.Schemas = schemas
+	p.TotalResults = totalResults
+	return &p
+}
+
+// NewGetSchemasResponseWithDefaults instantiates a new GetSchemasResponse object.
+// This constructor will only assign default values to properties that have it defined,
+// but it doesn't guarantee that properties required by API are set
+func NewGetSchemasResponseWithDefaults() *GetSchemasResponse {
 	p := GetSchemasResponse{}
 	return &p
 }
@@ -64,18 +74,19 @@ func (o *GetSchemasResponse) SetItemsPerPage(v int32) {
 	o.ItemsPerPage = &v
 }
 
-// GetSchemas returns the Schemas field value if set, zero value otherwise.
+// GetSchemas returns the Schemas field value.
 func (o *GetSchemasResponse) GetSchemas() []string {
-	if o == nil || o.Schemas == nil {
+	if o == nil {
 		var ret []string
 		return ret
 	}
-	return *o.Schemas
+
+	return o.Schemas
 }
 
-// SetSchemas gets a reference to the given []string and assigns it to the Schemas field.
+// SetSchemas sets field value.
 func (o *GetSchemasResponse) SetSchemas(v []string) {
-	o.Schemas = &v
+	o.Schemas = v
 }
 
 // GetStartIndex returns the StartIndex field value if set, zero value otherwise.
@@ -92,16 +103,17 @@ func (o *GetSchemasResponse) SetStartIndex(v int32) {
 	o.StartIndex = &v
 }
 
-// GetTotalResults returns the TotalResults field value if set, zero value otherwise.
+// GetTotalResults returns the TotalResults field value.
 func (o *GetSchemasResponse) GetTotalResults() int32 {
-	if o == nil || o.TotalResults == nil {
+	if o == nil {
 		var ret int32
 		return ret
 	}
-	return *o.TotalResults
+
+	return o.TotalResults
 }
 
-// SetTotalResults gets a reference to the given int32 and assigns it to the TotalResults field.
+// SetTotalResults sets field value.
 func (o *GetSchemasResponse) SetTotalResults(v int32) {
-	o.TotalResults = &v
+	o.TotalResults = v
 }